# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Robot Shop",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/robots": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "搜索时支持名称、描述和所有者筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Robot"
                ],
                "summary": "获取机器人列表",
                "operationId": "ListRobots",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "描述",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "所有者",
                        "name": "owner",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.RobotSearchResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一个新的机器人",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Robot"
                ],
                "summary": "创建机器人",
                "operationId": "RobotCreate",
                "parameters": [
                    {
                        "description": "机器人数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.RobotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/robots/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取指定ID的机器人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Robot"
                ],
                "summary": "获取机器人",
                "operationId": "GetRobot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "机器人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.RobotResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新机器人数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Robot"
                ],
                "summary": "更新机器人",
                "operationId": "RobotUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "机器人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "机器人数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.RobotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定ID的机器人",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Robot"
                ],
                "summary": "删除机器人",
                "operationId": "RobotDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "机器人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Robot": {
            "type": "object",
            "properties": {
                "callback": {
                    "description": "回调地址",
                    "type": "string",
                    "example": "https://example.com/callback"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "desc": {
                    "description": "描述",
                    "type": "string",
                    "example": "It's a robot"
                },
                "enabled": {
                    "description": "是否启用",
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "example": "bot"
                },
                "owner": {
                    "description": "所有者",
                    "type": "string",
                    "example": "Billy"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "webhook": {
                    "description": "通知地址",
                    "type": "string",
                    "example": "https://example.com/webhook"
                }
            }
        },
        "RobotList": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Robot"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
        "backend_api_v1.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.RobotRequest": {
            "type": "object",
            "properties": {
                "callback": {
                    "description": "回调地址",
                    "type": "string",
                    "example": "https://example.com/callback"
                },
                "desc": {
                    "description": "描述",
                    "type": "string",
                    "example": "It's a robot"
                },
                "enabled": {
                    "description": "是否启用",
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "example": "bot"
                },
                "owner": {
                    "description": "所有者",
                    "type": "string",
                    "example": "Billy"
                },
                "webhook": {
                    "description": "通知地址",
                    "type": "string",
                    "example": "https://example.com/webhook"
                }
            }
        },
        "backend_api_v1.RobotResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/Robot"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.RobotSearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/RobotList"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}